Move COMMA_SEPARATED_LIST_REGEX to constants
